requirements.txt
fink-client>=7.0
fastavro>=1.9
PyYAML>=6.0
requests>=2.31
tabulate>=0.9
pydantic>=2.7
python-dotenv>=1.0
